requirements.txt.txt
streamlit
requests
python-dotenv   # optional for local dev
